add box

+++ rust/std/macros.rs
@@ -4,7 +4,7 @@ macro_rules! pages {
 		use sys::getpagesize;
 		let size = unsafe { getpagesize() };
 		if size > 0 {
-			1 + ($v as u64 - 1 as u64) / size as u64
+			1 + ($v as usize - 1 as usize) / size as usize
 		} else {
 			0
 		}
@@ -16,7 +16,7 @@ macro_rules! page_size {
 	() => {{
 		#[allow(unused_unsafe)]
 		use sys::getpagesize;
-		let v = unsafe { getpagesize() } as u64;
+		let v = unsafe { getpagesize() } as usize;
 		v
 	}};
 }
