+++ rust/sys.rs
@@ -3,8 +3,8 @@ extern "C" {
 	pub fn read(fd: i32, buf: *mut u8, len: usize) -> i64;
 	pub fn write(fd: i32, buf: *const u8, len: usize) -> i64;
 	pub fn _exit(code: i32);
-	pub fn map(pages: u64) -> *mut u8;
-	pub fn unmap(ptr: *mut u8, pages: u64);
+	pub fn map(pages: usize) -> *mut u8;
+	pub fn unmap(ptr: *mut u8, pages: usize);
 	pub fn getpagesize() -> i32;
 }
